Consolidate messaging to utility function
Adding the event grid trigger requires reuse of same TG messaging functionality.

Add readme
