+++ ButtonToBot.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26124.0
 MinimumVisualStudioVersion = 15.0.26124.0
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DirectToTelegram", "DirectToTelegram\DirectToTelegram.csproj", "{EA6C68A2-C772-4D47-92BE-BD6574E4DA7E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DirectToTelegram", "DirectToTelegram\DirectToTelegram.csproj", "{EA6C68A2-C772-4D47-92BE-BD6574E4DA7E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
